+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Fixed a crash on launch when relay model was outdated.
 - Fix your own posts showing as outside your network on a fresh install. 
+- Add self-hosted PostHog analytics
+- Render mentions on notifications tab
 
 ## [0.1 (8)] - 2023-03-13Z
 - Fix translucent tab bar in the simulator.
@@ -23,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use only relays added in RelayView for sending and receiving events
 - Add PostHog analytics
 - Render note mentions in NoteCard
-- Add PostHog analytics
 - Open an email compose view controller for support@planetary.social 
 - Fix duplicate note on a new post
 - Add mute functionality